support virtio serial console support
It can be enabled by providing a file in the kernel package.
Create it via

  echo virtio > $RPM_BUILD_ROOT/.build.console.kvm

This needed some code cleanup to avoid breakages with qemu system
emulator. Started to move some common code of kvm and qemu into
vm_startup_qemu_params() method

This commit has currently a dependency to socat, so it can not get
merged yet.